added GetPopularDirectionById method

+++ src/main/java/com/example/avialine/repo/PopularDirectoryRepo.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PopularDirectoryRepo extends JpaRepository<com.example.avialine.model.entity.PopularDirectory, Long> {
+public interface PopularDirectoryRepo extends JpaRepository<com.example.avialine.model.entity.PopularDirectory, Integer> {
 }
